Stores encryption keys in a separate SQL database.

+++ PD/Controllers/UploadController.cs
@@ -16,10 +16,13 @@ namespace PD.Controllers
     {
         private readonly ApplicationDbContext _context;
         private readonly IDataProtectionProvider _provider;
-        public UploadController(ApplicationDbContext context, IDataProtectionProvider provider)
+        private readonly IConfiguration _configuration;
+
+        public UploadController(ApplicationDbContext context, IDataProtectionProvider provider, IConfiguration configuration)
         {
             _context = context;
             _provider = provider;
+            _configuration = configuration;
         }
 
         [HttpGet]
@@ -38,8 +41,9 @@ namespace PD.Controllers
             string path = Path.GetTempPath();
             string filenamePrefix = Path.GetRandomFileName();
 
-            foreach (var formFile in files)
+            if (files.Count() > 0)
             {
+                var formFile = files[0];
                 if (formFile.Length > 0)
                 {
                     string tmpFile = Path.Combine(path, filenamePrefix).Replace('.', '_') + "_" + formFile.FileName;
@@ -48,13 +52,35 @@ namespace PD.Controllers
                         await formFile.CopyToAsync(stream);
                     }
 
-                    //processing the uploaded file
-                    ImportService ds = new ImportService(_context);
-                    ds.InjestFacultySalaryAdjustmentData(tmpFile, worksheetName, new DateTime(2015, 7, 1).Date,
-                        new DateTime(2016,06,30).Date, new DateTime(2017, 06, 30).Date);
+                    BackgroundService srv = new BackgroundService(_context);
+                    string jobKey = "Faculty Data import";
+                    string jobId = srv.Enque(() => 
+                        new ImportService(new BackgroundService(_configuration).CreateDbContext())
+                            .InjestFacultySalaryAdjustmentData(tmpFile, worksheetName, new DateTime(2015, 7, 1).Date,
+                                new DateTime(2016, 06, 30).Date, new DateTime(2017, 06, 30).Date, true),
+                        jobKey);
 
-                    //Deleting the temporary file
-                    System.IO.File.Delete(tmpFile);
+                    if(string.IsNullOrEmpty(jobId))
+                    {
+                        ViewBag.Message = "Previous job is still running. No new job added.";
+                        ViewBag.MessageClass = "alert-danger";
+                        System.IO.File.Delete(tmpFile);
+                    }
+                    else
+                    {
+                        ViewBag.Message = string.Format("New job Added. Job ID: {0}", jobId);
+                        ViewBag.MessageClass = "alert-info";
+                    }
+                    ViewBag.FilePath = tmpFile;
+
+
+                    //////processing the uploaded file
+                    ////ImportService ds = new ImportService(_context);
+                    ////ds.InjestFacultySalaryAdjustmentData(tmpFile, worksheetName, new DateTime(2015, 7, 1).Date,
+                    ////    new DateTime(2016,06,30).Date, new DateTime(2017, 06, 30).Date);
+
+                    //////Deleting the temporary file
+                    ////System.IO.File.Delete(tmpFile);
                 }
             }
